requirements.txt
fastapi
uvicorn[standard]
celery
redis
litellm
requests
python-dotenv
pytest
requests-mock
streamlit